Fix to require BrowserWindow and ipcMain in function
