+++ src/components/retweet.js
@@ -12,6 +12,7 @@ export default class Retweet extends Tweet {
     tab:     PropTypes.string.isRequired,
     tweet:   PropTypes.object.isRequired,
     onClick: PropTypes.func.isRequired,
+    openMediaInWindow: PropTypes.func.isRequired,
   }
 
   render() {
