+++ src/containers/tweet-container.js
@@ -15,6 +15,45 @@ export default class TweetContainer extends BaseContainer {
     this.store.dispatch(Actions.selectTweet(this.props.tweet, this.props.tab, this.props.account));
   }
 
+  openMediaInWindow(media) {
+    const { BrowserWindow } = require('electron').remote;
+    const ipcMain           = require('electron').remote.ipcMain;
+
+    let options = { resizable: false, width: 100, height: 100 };
+    if (process.platform === 'darwin'){
+      Object.assign(options, { titleBarStyle: 'hidden' });
+    }
+    let win = new BrowserWindow(options);
+
+    win.loadURL(`file://${__dirname}../../media-popup.html`);
+
+    win.webContents.on('did-finish-load', () => {
+      let mediaUrl  = media.media_url;
+      let mediaType = '';
+      if (media.video_info != null) {
+        mediaUrl  = media.video_info.variants[0].url;
+        mediaType = media.video_info.variants[0].content_type;
+      }
+      win.webContents.send('load-media', mediaUrl, mediaType);
+    });
+
+    ipcMain.on('imageDimensions', (event, width, height) => {
+      if (win != null) {
+        const screenWidth  = Math.round((screen.width/2) - (width/2));
+        const screenHeight = Math.round((screen.height/3) - (height/3));
+        let options = { x: screenWidth, y: screenHeight, width: width, height: height }
+        if (process.platform === 'darwin') {
+          Object.assign(options, { height: height+20 });
+        }
+        win.setBounds(options, true);
+      }
+    });
+
+    win.on('closed', () => {
+      win = null;
+    });
+  }
+
   render() {
     if (this.props.tweet.retweeted_status) {
       return(
@@ -24,6 +63,7 @@ export default class TweetContainer extends BaseContainer {
           tab={this.props.tab}
           tweet={this.props.tweet}
           onClick={this.onClick.bind(this)}
+          openMediaInWindow={this.openMediaInWindow.bind(this)}
           now={this.props.now}
         />
       );
@@ -35,6 +75,7 @@ export default class TweetContainer extends BaseContainer {
           tab={this.props.tab}
           tweet={this.props.tweet}
           onClick={this.onClick.bind(this)}
+          openMediaInWindow={this.openMediaInWindow.bind(this)}
           now={this.props.now}
         />
       );
